refactor utils/types
