+++ ui/components/src/Table/table-model.ts
@@ -1,0 +1,1 @@
+export type TableDensity = 'compact' | 'standard';
